backend/requirements.txt
crewai==0.83.1
crewai-tools==0.20.0
langchain==0.3.14
langchain-openai==0.2.14
python-dotenv==1.0.1
portkey-ai==1.8.2
